backup/requirements.txt
tqdm
python-dotenv
Pillow
psycopg2-binary
SQLAlchemy
azure-storage-blob
